test: add tests for AppDrawerPagePresenter
