+++ feature/appdrawerpage/src/main/kotlin/dev/mslalith/focuslauncher/feature/appdrawerpage/AppDrawerPage.kt
@@ -9,6 +9,7 @@ import androidx.compose.foundation.layout.imePadding
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
@@ -19,8 +20,10 @@ import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.slack.circuit.codegen.annotations.CircuitInject
+import com.slack.circuit.overlay.LocalOverlayHost
 import com.slack.circuit.runtime.screen.Screen
 import dagger.hilt.components.SingletonComponent
+import dev.mslalith.focuslauncher.core.circuitoverlay.showBottomSheet
 import dev.mslalith.focuslauncher.core.common.extensions.launchApp
 import dev.mslalith.focuslauncher.core.common.model.LoadingState
 import dev.mslalith.focuslauncher.core.model.AppDrawerViewType
@@ -35,6 +38,7 @@ import dev.mslalith.focuslauncher.core.ui.providers.LocalLauncherPagerState
 import dev.mslalith.focuslauncher.feature.appdrawerpage.apps.grid.AppsGrid
 import dev.mslalith.focuslauncher.feature.appdrawerpage.apps.list.AppsList
 import kotlinx.coroutines.flow.collectLatest
+import kotlinx.coroutines.launch
 
 @CircuitInject(AppDrawerPageScreen::class, SingletonComponent::class)
 @Composable
@@ -46,14 +50,19 @@ fun AppDrawerPage(
     // treats it as stable. See: https://issuetracker.google.com/issues/256100927
     val eventSink = state.eventSink
 
-    fun openBottomSheet(screen: Screen) = eventSink(AppDrawerPageUiEvent.OpenBottomSheet(screen = screen))
+    val scope = rememberCoroutineScope()
+    val overlayHost = LocalOverlayHost.current
+
+    fun showBottomSheet(screen: Screen) {
+        scope.launch { overlayHost.showBottomSheet(screen) }
+    }
 
     AppDrawerPageKeyboardAware(
         modifier = modifier,
         state = state,
         onSearchQueryChange = { eventSink(AppDrawerPageUiEvent.UpdateSearchQuery(query = it)) },
         reloadIconPack = { eventSink(AppDrawerPageUiEvent.ReloadIconPack) },
-        showAppMoreOptions = { openBottomSheet(screen = AppMoreOptionsBottomSheetScreen(appDrawerItem = it)) }
+        showAppMoreOptions = { showBottomSheet(screen = AppMoreOptionsBottomSheetScreen(appDrawerItem = it)) }
     )
 }
