+++ feature/appdrawerpage/src/main/kotlin/dev/mslalith/focuslauncher/feature/appdrawerpage/AppDrawerPagePresenter.kt
@@ -9,11 +9,8 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.runtime.snapshotFlow
 import com.slack.circuit.codegen.annotations.CircuitInject
-import com.slack.circuit.overlay.LocalOverlayHost
 import com.slack.circuit.runtime.presenter.Presenter
-import com.slack.circuit.runtime.screen.Screen
 import dagger.hilt.components.SingletonComponent
-import dev.mslalith.focuslauncher.core.circuitoverlay.showBottomSheet
 import dev.mslalith.focuslauncher.core.common.appcoroutinedispatcher.AppCoroutineDispatcher
 import dev.mslalith.focuslauncher.core.common.model.LoadingState
 import dev.mslalith.focuslauncher.core.data.repository.settings.AppDrawerSettingsRepo
@@ -26,7 +23,6 @@ import dev.mslalith.focuslauncher.core.model.Constants.Defaults.Settings.AppDraw
 import dev.mslalith.focuslauncher.core.model.Constants.Defaults.Settings.AppDrawer.DEFAULT_SEARCH_BAR
 import dev.mslalith.focuslauncher.core.model.appdrawer.AppDrawerItem
 import dev.mslalith.focuslauncher.core.screens.AppDrawerPageScreen
-import dev.mslalith.focuslauncher.feature.appdrawerpage.AppDrawerPageUiEvent.OpenBottomSheet
 import dev.mslalith.focuslauncher.feature.appdrawerpage.AppDrawerPageUiEvent.ReloadIconPack
 import dev.mslalith.focuslauncher.feature.appdrawerpage.AppDrawerPageUiEvent.UpdateSearchQuery
 import kotlinx.collections.immutable.ImmutableList
@@ -59,7 +55,6 @@ class AppDrawerPagePresenter @Inject constructor(
     @Composable
     override fun present(): AppDrawerPageState {
         val scope = rememberCoroutineScope()
-        val overlayHost = LocalOverlayHost.current
 
         val appDrawerViewType by appDrawerSettingsRepo.appDrawerViewTypeFlow.collectAsState(initial = DEFAULT_APP_DRAWER_VIEW_TYPE)
         val appDrawerIconViewType by appDrawerSettingsRepo.appDrawerIconViewTypeFlow.collectAsState(initial = DEFAULT_APP_DRAWER_ICON_VIEW_TYPE)
@@ -81,10 +76,6 @@ class AppDrawerPagePresenter @Inject constructor(
                 .collect()
         }
 
-        fun showBottomSheet(screen: Screen) {
-            scope.launch { overlayHost.showBottomSheet(screen) }
-        }
-
         return AppDrawerPageState(
             allAppsState = allAppsState,
             appDrawerViewType = appDrawerViewType,
@@ -96,7 +87,6 @@ class AppDrawerPagePresenter @Inject constructor(
             when (it) {
                 is UpdateSearchQuery -> searchBarQuery = it.query
                 ReloadIconPack -> scope.reloadIconPack()
-                is OpenBottomSheet -> showBottomSheet(screen = it.screen)
             }
         }
     }
