+++ feature/appdrawerpage/src/main/kotlin/dev/mslalith/focuslauncher/feature/appdrawerpage/AppDrawerPageUiEvent.kt
@@ -1,10 +1,8 @@
 package dev.mslalith.focuslauncher.feature.appdrawerpage
 
 import com.slack.circuit.runtime.CircuitUiEvent
-import com.slack.circuit.runtime.screen.Screen
 
 sealed interface AppDrawerPageUiEvent : CircuitUiEvent {
     data class UpdateSearchQuery(val query: String) : AppDrawerPageUiEvent
     data object ReloadIconPack : AppDrawerPageUiEvent
-    data class OpenBottomSheet(val screen: Screen) : AppDrawerPageUiEvent
 }
